==> 2. Feature Engineering.ipynb <==
import pickle
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
from sklearn.tree import DecisionTreeRegressor
from sklearn.ensemble import RandomForestRegressor
from sklearn.model_selection import train_test_split
from sklearn.metrics import mean_squared_error
from sklearn.base import clone
from sklearn.tree import DecisionTreeClassifier

import pandas as pd
import numpy as np
import os

pd.set_option('display.max_columns', None)

data=pd.read_csv("ObesityDataSet.csv")

data.head()

data.info()

data.describe()

data["NObeyesdad"].value_counts()/len(data)*100

### Descripción de columnas utiles

Basados en el analisis de la data que se realizó, se puede observar que el campo de edad presenta rangos utiles que pueden ser combinados con otros campos para mejorar la performance del modelo. El campo de family_history_with_overweight presenta una correlacion de 0.51 con el target y en el analisis se demostró que para un BMI mayor a 28 (sobrepeso-obesidad) hay una tendencia a tener familiares con sobrepeso.

Por otro lado, BMI es un indicador que se elaboró que demostro una correlacion muy alta con los niveles de obesidad y tiene rangos bien delimitados por categorias. Esto se puede apreciar visualmente en el analisis de la data. Por esta razon, no se realizara algun binning o trabajo con esta columna.

Otros campos numericos con los que se podria realizar alguna combinacion son : Frequent consumption of high caloric food (FAVC) y Physical activity frequency (FAF). Estos campos tienen una buena correlacion connel target y adicionalmente de manera empirica podemos afirmar que son factores importantes para poder predecir el nivel de obesidad. En el analisis de la data FAVC demostro un pico en 1.00 en donde todas las categorias se reunen con diferentes intensidades, por lo que no se puede aprovechar  de hacer algun tipo de binning en esta columna. Por otro lado, FAF demostro tener unos rangos utiles cercanos al 1,2 y 3.


### Proposito del Feature Engineering

Este dataset no tiene datos nulos o categorias con un valor desconocido. El target esta bien distribuido por las categorias. Por lo tanto, el siguiente paso seria cambiar las columnas ordinales por numeros que demuestren la jerarquia de las opciones y las columnas binarias por valores de 0 y 1. Por ultimo, transformar a One Hot encoding las columnas categoricas. En este caso solo MTRANS.

Ademas, lo que se busca encontrar en los proyectos de data science son grupos especificos que puedan ayudar al modelo a identificar clases. Por esta razon, se plantean combinaciones de columnas para luego realizar One Hot Encoding. La utilidad de estas columnas asi como un Feature Selection, de ser necesario, sera evaluado en otros notebooks.

### Combinaciones propuestas:
    - Gender | Age (Bin)
    - CAEC | Age (Bin)
    - Gender | family_history_with_overweight
    - Gender | FAF(Bin)

data.head()

### Binning de Age y FAF

Los bins de edad de localizaran segun el analisis de la data. Los grupos mas importantes son de [15-20], [24-30] y [30-35]. Se realizaran particiones en 15, 20, 24, 30 y 35. Menores de 15 solo hay dos personas por lo que se adaptara al rango de [0-20]

cut_labels = [1, 2, 3, 4,5]
cut_bins = [0, 20, 24, 30,35,999]
data['Age_bin'] = pd.cut(data['Age'], bins=cut_bins, labels=cut_labels)

data['Age_bin'].value_counts()

data[data['Age_bin'].isnull()]

Los bins de FAF se localizaran segun el analisis de la data. Los grupos mas importantes son de [0-0.5], [0.5-1.5] y [1.5-2.5]. Se realizaran particiones en 0.5, 1.5, 2.5.

cut_labels = [1, 2, 3, 4]
cut_bins = [-1,0.5, 1.5, 2.5, 99999]
data['FAF_bin'] = pd.cut(data['FAF'], bins=cut_bins, labels=cut_labels)

data['FAF_bin'].value_counts()

data[data['FAF_bin'].isnull()]

Se confirma que los bins han sido elaborados correctamente por lo que se pasa a la commbinación de columnas.

### Combinacion de columnas

data["Gender|Age"]=data["Gender"]+"_" +data["Age_bin"].astype(str)

data["CAEC|Age"]=data["CAEC"]+"_" +data["Age_bin"].astype(str)

data["Gender|FO"]=data["Gender"]+"_" +data["family_history_with_overweight"]

data["Gender|FAF"]=data["Gender"]+"_" +data["FAF_bin"].astype(str)

### Transformando columnas binarias,categoricas y ordinales

numeric=["Age","Height","Weight","FCVC","NCP","CH2O","FAF","TUE"]
ordinal=["CAEC","CALC"]
flags=["family_history_with_overweight","FAVC","SMOKE","SCC"]
categorical=["MTRANS"]

data["CAEC"]=data["CAEC"].map({"no":0,"Sometimes":1,"Frequently":2,"Always":3})

data["CALC"]=data["CALC"].map({"no":0,"Sometimes":1,"Frequently":2,"Always":3})

data["Gender"]=data["Gender"].map({"Female":0,"Male":1})

for column in flags:
    data[column]=data[column].map({"yes":1,"no":0})

data["MTRANS"] = data["MTRANS"].astype("category")

dictionary_of_columns_with_index_to_categorical = dict()
dictionary_of_columns_with_categorical_to_index = dict()
    
dictionary_of_columns_with_index_to_categorical["MTRANS"] = dict()
dictionary_of_columns_with_categorical_to_index["MTRANS"] = dict()

for index, category in enumerate(data["MTRANS"].cat.categories):
        dictionary_of_columns_with_index_to_categorical["MTRANS"][index] = category
        dictionary_of_columns_with_categorical_to_index["MTRANS"][category] = index
        
data.replace(dictionary_of_columns_with_categorical_to_index, inplace=True)

data["NObeyesdad"]=data["NObeyesdad"].map({"Insufficient_Weight":0,"Normal_Weight":1,"Overweight_Level_I":2,
                                           "Overweight_Level_II":3,"Obesity_Type_I":4,"Obesity_Type_II":5,
                                           "Obesity_Type_III":6})

### Se crea la columna BMI 

data["BMI"]=data["Weight"]/(data["Height"]*data["Height"])

### Ordenamiento de columnas y split de los datos

Se crearan dos versiones del dataset. La primera version es el dataset base procesando las columnas, y agregandole el BMI. La segunda version es la que utiliza Age_bin y FAF_bin y realiza feature engineering. Los dataset se pueden variar un poco para el entrenamiento de los modelos para mejorar su performance

#### Particion segunda version (Combinacion de columnas y binning)

data_2=pd.get_dummies(data,columns=['MTRANS', 'Gender|Age', 'CAEC|Age',
                           'Gender|FO','Gender|FAF'])

data_2.columns

data_2=data_2[['Gender', 'Age', 'Height', 'Weight', 'BMI','family_history_with_overweight',
       'FAVC', 'FCVC', 'NCP', 'CAEC', 'SMOKE', 'CH2O', 'SCC', 'FAF', 'TUE',
       'CALC', 'NObeyesdad', 'Age_bin', 'FAF_bin','MTRANS_0',
       'MTRANS_1', 'MTRANS_2', 'MTRANS_3', 'MTRANS_4', 'Gender|Age_Female_1',
       'Gender|Age_Female_2', 'Gender|Age_Female_3', 'Gender|Age_Female_4',
       'Gender|Age_Female_5', 'Gender|Age_Male_1', 'Gender|Age_Male_2',
       'Gender|Age_Male_3', 'Gender|Age_Male_4', 'Gender|Age_Male_5',
       'CAEC|Age_Always_1', 'CAEC|Age_Always_2', 'CAEC|Age_Always_3',
       'CAEC|Age_Always_4', 'CAEC|Age_Always_5', 'CAEC|Age_Frequently_1',
       'CAEC|Age_Frequently_2', 'CAEC|Age_Frequently_3',
       'CAEC|Age_Frequently_4', 'CAEC|Age_Frequently_5',
       'CAEC|Age_Sometimes_1', 'CAEC|Age_Sometimes_2', 'CAEC|Age_Sometimes_3',
       'CAEC|Age_Sometimes_4', 'CAEC|Age_Sometimes_5', 'CAEC|Age_no_1',
       'CAEC|Age_no_2', 'CAEC|Age_no_3', 'CAEC|Age_no_4',
       'Gender|FO_Female_no', 'Gender|FO_Female_yes', 'Gender|FO_Male_no',
       'Gender|FO_Male_yes', 'Gender|FAF_Female_1', 'Gender|FAF_Female_2',
       'Gender|FAF_Female_3', 'Gender|FAF_Female_4', 'Gender|FAF_Male_1',
       'Gender|FAF_Male_2', 'Gender|FAF_Male_3', 'Gender|FAF_Male_4']]

data_2.head()

x_2 = data_2.drop("NObeyesdad", axis=1)
y_2 = data_2["NObeyesdad"]

x_train_2, x_test_2, y_train_2, y_test_2 = train_test_split(x_2, y_2, test_size=0.2, random_state=42,stratify=y)

print("Tamaño de la data de entrenamiento:{}".format(len(x_train_2)))

print("Tamaño de la data de entrenamiento:{}".format(len(y_test_2)))

x_train_2.head()

x_train_2.to_csv("X_train_v2.csv",index=False)
y_train_2.to_csv("Y_train_v2.csv",index=False)
x_test_2.to_csv("X_test_v2.csv",index=False)
y_test_2.to_csv("Y_test_v2.csv",index=False)

#### Particion primera version (Solo BMI extra)

data.columns

data=data[['Gender', 'Age', 'Height', 'Weight','BMI', 'family_history_with_overweight',
       'FAVC', 'FCVC', 'NCP', 'CAEC', 'SMOKE', 'CH2O', 'SCC', 'FAF', 'TUE',
       'CALC', 'MTRANS', 'NObeyesdad']]

data.head()

x = data.drop("NObeyesdad", axis=1)
y = data["NObeyesdad"]

x_train, x_test, y_train, y_test = train_test_split(x, y, test_size=0.2, random_state=42,stratify=y)

print("Tamaño de la data de entrenamiento:{}".format(len(x_train)))

print("Tamaño de la data de entrenamiento:{}".format(len(y_test)))

x_train.head()

x_train.to_csv("X_train.csv",index=False)
y_train.to_csv("Y_train.csv",index=False)
x_test.to_csv("X_test.csv",index=False)
y_test.to_csv("Y_test.csv",index=False)
