==> 3. Recreación del Baseline.ipynb <==
import pandas as pd
import numpy as np
import os

from sklearn.tree import DecisionTreeRegressor
from sklearn.model_selection import cross_val_score
from sklearn.model_selection import GridSearchCV
from sklearn.model_selection import train_test_split
from sklearn.metrics import classification_report
from sklearn.metrics import mean_squared_error
from sklearn.base import clone
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import confusion_matrix, classification_report 
from sklearn.metrics import accuracy_score
from sklearn.metrics import classification_report
from sklearn.metrics import roc_curve
from sklearn.metrics import roc_auc_score

### Leyendo dataset base

x_train=pd.read_csv("X_train.csv")
x_test=pd.read_csv("X_test.csv")
y_train=pd.read_csv("Y_train.csv")
y_test=pd.read_csv("Y_test.csv")

x_train.head()

x_train.drop(columns=["BMI"],inplace=True)

x_test.drop(columns=["BMI"],inplace=True)

len(x_train),len(x_test)

### Entrenamos con un RF rapidamente para analizar si se optiene el mismo baseline

rf = RandomForestClassifier(criterion='entropy', max_depth=20, n_estimators=2600, oob_score=True, class_weight = 'balanced', random_state=1, max_features='sqrt')
rf.fit(x_train, y_train)

rf.score(x_train, y_train)

rf.score(x_test, y_test)

### Como en el paper no se menciona el tamaño del test no podemos tener un punto de comparacion igual pero se aplicara GridSearchCV  para obtener el mejor resultado usando CV

No usamos oob score para usar toda la data en el entrenamiento

#Entrenamos el modelo random forest de clasificación usando el GridSearchCV para considerar los mejores hiperparámetros
rf= RandomForestClassifier(random_state=42)

# grid searh to choose the best (combination of) hyperparameters, #'criterion': ['gini', 'entropy'], #'oob_score': [True, False], 'max_features':['sqrt', 0.1]

pg_rf={'criterion': ['gini', 'entropy'],'max_depth': [18, 20, 22], 'n_estimators': [2000,2600,2800], 
       'oob_score': [False], 'min_samples_split':[5,7] }

gs_rf= GridSearchCV (estimator= rf,
               param_grid= pg_rf,
               scoring='f1_macro',
               cv=5,refit=True)

# nested cross validation combining grid search (inner loop) and k-fold cv (outter loop)
gs_rf_scores = cross_val_score(gs_rf, X=x_train, y=y_train, cv=5, scoring='f1_macro', verbose = 10, n_jobs=-1)

# fit, and fit with best estimator
gs_rf.fit(x_train, y_train)

print('Train Accuracy:   {0:.1f}%'.format(gs_rf.score(x_train, y_train)*100))
print('CV Mean Accuracy: {0:.1f}%'.format(np.mean(gs_rf_scores)*100))
print('Test Accuracy:    {0:.1f}%'.format(gs_rf.score(x_test, y_test)*100))

Train Accuracy:   99.9%
CV Mean Accuracy: 94.5%
Test Accuracy:    94.7%

gs_rf.best_estimator_

rf= DecisionTreeClassifier(random_state=42,class_weight= "balanced")

# grid searh to choose the best (combination of) hyperparameters, #'criterion': ['gini', 'entropy'], #'oob_score': [True, False], 'max_features':['sqrt', 0.1]

pg_rf={'criterion': ['gini', 'entropy'],'max_depth': [6,8,10,12,16], 'min_samples_split':[2,3,4,6,7] }

gs_rf= GridSearchCV (estimator= rf,
               param_grid= pg_rf,
               scoring='f1_macro',
               cv=5,refit=True)

gs_rf.fit(x_train, y_train)

gs_rf_scores = cross_val_score(gs_rf, X=x_train, y=y_train, cv=5, scoring='f1_macro', verbose = 10, n_jobs=-1)

print('Train Accuracy:   {0:.1f}%'.format(gs_rf.score(x_train, y_train)*100))
print('CV Mean Accuracy: {0:.1f}%'.format(np.mean(gs_rf_scores)*100))
print('Test Accuracy:    {0:.1f}%'.format(gs_rf.score(x_test, y_test)*100))

gs_rf.best_params_

print(classification_report(list(y_test["NObeyesdad"]),gs_rf.predict(x_test)))
