==> 7. AdaBoost y CatBoost.ipynb <==
from sklearn.tree import DecisionTreeRegressor
from sklearn.model_selection import cross_val_score
from sklearn.model_selection import GridSearchCV
from sklearn.model_selection import train_test_split
from sklearn.metrics import classification_report
from sklearn.metrics import mean_squared_error
from sklearn.base import clone
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import confusion_matrix, classification_report 
from sklearn.metrics import accuracy_score
from sklearn.metrics import classification_report
from sklearn.metrics import roc_curve
from sklearn.metrics import roc_auc_score
import pandas as pd
import numpy as np
import os
from sklearn.svm import SVC

x_train=pd.read_csv("X_train.csv")
x_test=pd.read_csv("X_test.csv")
y_train=pd.read_csv("Y_train.csv")
y_test=pd.read_csv("Y_test.csv")

x_train_2=pd.read_csv("X_train_v2.csv")
x_test_2=pd.read_csv("X_test_v2.csv")
y_train_2=pd.read_csv("Y_train_v2.csv")
y_test_2=pd.read_csv("Y_test_v2.csv")

### Entrenando a AdaBoost con el modelo por deault

clf = AdaBoostClassifier(n_estimators=100, random_state=42)
model=clf.fit(x_train, y_train)

model.score(x_train, y_train)

### Entrenando a AdaBoost con SVC

svc=SVC(probability=True, kernel='linear')

# Create adaboost classifer object
abc =AdaBoostClassifier(n_estimators=100, base_estimator=svc,learning_rate=1)

model = abc.fit(x_train, y_train)

model.score(x_train, y_train)

### Entrenando a AdaBoost con DT cambiando algunos parametros

Despues de un prueba y error con diferentes parametros se utilizo el max_depth = 18 que se hallo como mejor depth en el GridSearch del modelo baseline

DTC = DecisionTreeClassifier(random_state = 42, class_weight = "balanced",max_depth = 18)

ABC = AdaBoostClassifier(base_estimator = DTC)
model = ABC.fit(x_train, y_train)
print(model.score(x_train, y_train))
print(model.score(x_test, y_test))

#### Aplicando lo mismo con la version con feature engineering

DTC = DecisionTreeClassifier(random_state = 42, class_weight = "balanced",max_depth = 18)

ABC = AdaBoostClassifier(base_estimator = DTC)
model = ABC.fit(x_train_2, y_train_2)
print(model.score(x_train_2, y_train_2))
print(model.score(x_test_2, y_test_2))

### Entrenando a CatBoost

from catboost import Pool,CatBoostClassifier

train_dataset = Pool(data=x_train,label=y_train) 

test_dataset = Pool(data=x_test,label=y_test) 

model_catboost = CatBoostClassifier(learning_rate=0.1,iterations=200,
                           depth=12)

model_catboost.fit(train_dataset)

model_catboost.score(train_dataset)

model_catboost.score(test_dataset)

#### Aplicando lo mismo con la version con feature engineering

train_dataset = Pool(data=x_train_2,label=y_train_2) 

test_dataset = Pool(data=x_test_2,label=y_test_2) 

model_catboost = CatBoostClassifier(learning_rate=0.1,iterations=250,
                           depth=12)

model_catboost.fit(train_dataset)

model_catboost.score(train_dataset)

model_catboost.score(test_dataset)
